+++ tzarc-djinn/drivers/drawable/drawable.h
@@ -28,7 +28,7 @@ typedef void (*drawable_init_func)(drawable_driver_t *driver, drawable_rotation_
 typedef void (*drawable_clear_func)(drawable_driver_t *driver);
 typedef void (*drawable_power_func)(drawable_driver_t *driver, bool power_on);
 typedef void (*drawable_viewport_func)(drawable_driver_t *driver, uint16_t left, uint16_t top, uint16_t right, uint16_t bottom);
-typedef void (*drawable_pixdata_func)(drawable_driver_t *driver, const uint16_t *pixel_data, uint32_t num_pixels);
+typedef void (*drawable_pixdata_func)(drawable_driver_t *driver, const uint8_t *pixel_data, uint32_t num_pixels);
 typedef void (*drawable_setpixel_func)(drawable_driver_t *driver, uint16_t x, uint16_t y, uint16_t color);
 typedef void (*drawable_line_func)(drawable_driver_t *driver, uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1, uint16_t color);
 typedef void (*drawable_rect_func)(drawable_driver_t *driver, uint16_t left, uint16_t top, uint16_t right, uint16_t bottom, uint16_t color, bool filled);
@@ -49,7 +49,7 @@ void drawable_init(drawable_driver_t *driver, drawable_rotation_t rotation);
 void drawable_clear(drawable_driver_t *driver);
 void drawable_power(drawable_driver_t *driver, bool power_on);
 void drawable_viewport(drawable_driver_t *driver, uint16_t left, uint16_t top, uint16_t right, uint16_t bottom);
-void drawable_pixdata(drawable_driver_t *driver, const uint16_t *pixel_data, uint32_t num_pixels);
+void drawable_pixdata(drawable_driver_t *driver, const uint8_t *pixel_data, uint32_t num_pixels);
 void drawable_setpixel(drawable_driver_t *driver, uint16_t x, uint16_t y, uint16_t color);
 void drawable_line(drawable_driver_t *driver, uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1, uint16_t color);
 void drawable_rect(drawable_driver_t *driver, uint16_t left, uint16_t top, uint16_t right, uint16_t bottom, uint16_t color, bool filled);
